some fixes for issues found while playing hl2
fixed chaos bar flickering and being 1 pixel wide when chaos 0
made clientscheme.res match with 2013 SDK except in specifically needed cases
fixed intro credits text being too small
fixed gman being below the deck in canals 6 - this is something that went wrong in CAI_BaseNPC, but it seems to have only manifested as a problem in this single instance. it's not clear why this specific case has this problem. effect on any other NPCs is thought to be minimal. if it affects any other NPCs, they would have to be close to a ceiling.
fixed suppression device being unable to damage scripted targets
fixed camera in breen not looking good with ep2 advisor model, new model is because the cables would not appear to move when viewed on the monitor only